+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [3.0.0] 2021-01-28
+### Added
+- Updated URI used for score calculation extensions from:
+  `http://lhcforms.nlm.nih.gov/fhir/ext/rule-editor-expression` to
+  `http://lhcforms.nlm.nih.gov/fhir/ext/rule-editor-score-expression`
+- Score calculation logic for nested items:
+  - Assume scored items are above (in question order) the total score item.
+  - If a preceding item is also a total score item, don’t consider any earlier
+    items.
+
 ## [2.0.1] 2022-01-28
 ### Changed
 - Hide instant preview in case statements when condition, output and default is